Refactor to use toast notification component, errorpage fixed, account delete moved to my account
