+++ mycodereactserver/src/Pages/Login.jsx
@@ -3,13 +3,12 @@ import { useNavigate } from "react-router-dom";
 import { postApi } from "../Services/Api";
 import { getToken } from "../Services/AuthService";
 import { userLogin } from "../Services/Backend.Endpoints";
-import { toast } from "react-toastify";
 import { jwtDecode } from "jwt-decode";
 import Login from "../Components/Login/Login";
 import Loading from "../Components/Loading/Loading";
+import Notify from "./Services/ToastNotifications";
 import ErrorPage from "./Services/ErrorPage";
 import Cookies from "js-cookie";
-import "react-toastify/dist/ReactToastify.css";
 
 const UserLogin = () => {
     const navigate = useNavigate();
@@ -28,27 +27,9 @@ const UserLogin = () => {
                     Cookies.set("jwtToken", data.token, { expires: new Date(expirationTime) });
 
                     navigate("/");
-                    toast.success("Successful Login!", {
-                        position: "top-right",
-                        autoClose: 5000,
-                        hideProgressBar: false,
-                        closeOnClick: true,
-                        pauseOnHover: true,
-                        draggable: true,
-                        progress: undefined,
-                        theme: "dark",
-                    });
+                    Notify("Success", "Successful Login!");
                 } else {
-                    toast.error("Unable to Login!", {
-                        position: "top-right",
-                        autoClose: 5000,
-                        hideProgressBar: false,
-                        closeOnClick: true,
-                        pauseOnHover: true,
-                        draggable: true,
-                        progress: undefined,
-                        theme: "dark"
-                    });
+                    Notify("Error", "Unable to Login!");
                 }
             })
             .catch((error) => {
